+++ packages/client/src/index.ts
@@ -10,29 +10,7 @@ import { deserialize, serialize } from "@quatico/magellan-shared";
 import { remoteInvoke } from "./remote-invoke";
 import { Sdk } from "./sdk";
 import type { Configuration, Context } from "./transport";
-import {
-    addNamespace,
-    addNamespaceIfAbsent,
-    addTransport,
-    addTransportIfAbsent,
-    initProjectConfiguration,
-    resolveNamespace,
-    setNamespace,
-    setTransport
-} from "./transport";
+import { initProjectConfiguration, resolveNamespace, setNamespace, setTransport } from "./transport";
 
-export {
-    addNamespace,
-    addNamespaceIfAbsent,
-    addTransport,
-    addTransportIfAbsent,
-    deserialize,
-    initProjectConfiguration,
-    remoteInvoke,
-    resolveNamespace,
-    Sdk,
-    serialize,
-    setNamespace,
-    setTransport,
-};
+export { setNamespace, setTransport, deserialize, initProjectConfiguration, remoteInvoke, resolveNamespace, Sdk, serialize };
 export type { Configuration, Context, NamespaceMapping, TransportHandler };