+++ packages/client/src/transport/index.ts
@@ -8,6 +8,6 @@ export type { Configuration } from "./Configuration";
 export { getConfiguration, initProjectConfiguration } from "./configuration-repository";
 export type { Context } from "./Context";
 export { formdataFetch } from "./formdata-fetch";
-export { addNamespace, addNamespaceIfAbsent, addTransport, addTransportIfAbsent, resolveNamespace, setNamespace, setTransport } from "./namespace";
+export { setNamespace, setTransport, resolveNamespace } from "./namespace";
 export type { ResolvedNamespace } from "./ResolvedNamespace";
 export { transportRequest } from "./transport-request";