+++ packages/server/src/configuration/index.ts
@@ -5,20 +5,9 @@
  * ---------------------------------------------------------------------------------------------
  */
 import type { Configuration } from "./Configuration";
-import { applyExecutionContext, getConfiguration, initProjectConfiguration } from "./configuration-repository";
-import { addNamespace, addNamespaceIfAbsent, addTransport, addTransportIfAbsent, resolveNamespace, setNamespace, setTransport } from "./namespace";
+import { getConfiguration, initProjectConfiguration } from "./configuration-repository";
+import { resolveNamespace, setNamespace, setTransport } from "./namespace";
 import type { ResolvedNamespace } from "./ResolvedNamespace";
 
 export type { Configuration, ResolvedNamespace };
-export {
-    applyExecutionContext,
-    getConfiguration,
-    initProjectConfiguration,
-    addNamespace,
-    addNamespaceIfAbsent,
-    addTransport,
-    addTransportIfAbsent,
-    resolveNamespace,
-    setNamespace,
-    setTransport,
-};
+export { getConfiguration, initProjectConfiguration, setNamespace, setTransport, resolveNamespace };