+++ packages/server/src/index.ts
@@ -7,17 +7,7 @@
 import type { NamespaceMapping, TransportHandler } from "@quatico/magellan-shared";
 import type { Context } from "./api";
 import type { Configuration } from "./configuration";
-import {
-    addNamespace,
-    addNamespaceIfAbsent,
-    addTransport,
-    addTransportIfAbsent,
-    applyExecutionContext,
-    initProjectConfiguration,
-    resolveNamespace,
-    setNamespace,
-    setTransport
-} from "./configuration";
+import { initProjectConfiguration, resolveNamespace, setNamespace, setTransport } from "./configuration";
 import type { ServerOptions } from "./server";
 import { configureMagellanRoutes, configureRequestMiddlewares, serve, setupMagellanModules, startServer } from "./server";
 import { getFunctionService } from "./services";
@@ -25,11 +15,6 @@ import { formdataFetch } from "./transport";
 
 export type { ServerOptions, Configuration, Context, NamespaceMapping, TransportHandler };
 export {
-    addNamespace,
-    addNamespaceIfAbsent,
-    addTransport,
-    addTransportIfAbsent,
-    applyExecutionContext,
     configureMagellanRoutes,
     configureRequestMiddlewares,
     formdataFetch,