+++ packages/server/src/transport/formdata-fetch.spec.ts
@@ -1,7 +1,7 @@
 import { TransportFunction } from "@quatico/magellan-shared";
 import fetch from "node-fetch";
 import { initDependencyContext } from "../services";
-import { formdataFetch } from "./formdata-fetch";
+import { completeEndpoint, formdataFetch } from "./formdata-fetch";
 
 jest.mock("node-fetch");
 
@@ -19,7 +19,7 @@ describe("formdataFetch", () => {
 
         await formdataFetch(validTransportFn, { headers: {} });
 
-        expect(fetch).toHaveBeenCalledWith("http://localhost:3000//expected", expect.objectContaining({ method: "POST" }));
+        expect(fetch).toHaveBeenCalledWith("http://localhost:3000/expected", expect.objectContaining({ method: "POST" }));
         expect(fetch).toHaveBeenCalledTimes(1);
     });
 
@@ -79,5 +79,36 @@ describe("formdataFetch", () => {
             status: 404,
             message: "Expected Status Message",
         });
+    });
+});
+
+describe("completeEndpoint", () => {
+    it("called with endpoint that is a valid absolute url returns expected", () => {
+        const actual = completeEndpoint("https://www.valid.url");
+
+        expect(actual).toBe("https://www.valid.url");
+    });
+    it("called with endpoint that is a relative url with leading slash returns expected", () => {
+        const actual = completeEndpoint("/api");
+
+        expect(actual).toBe("http://localhost:3000/api");
+    });
+    it("called with endpoint that is an invalid absolute url (invalid protocol) throws", () => {
+        expect(() => completeEndpoint("hrrps://www.invalid.url")).toThrow("provided endpoint hrrps://www.invalid.url is invalid");
+    });
+    it("called with endpoint that is an invalid absolute url (no protocol) throws", () => {
+        expect(() => completeEndpoint("www.no-protocol.url")).toThrow("provided endpoint www.no-protocol.url is invalid");
+    });
+    it("called with endpoint that is an invalid absolute url (invalid character) throws", () => {
+        expect(() => completeEndpoint("http://www.with space.url")).toThrow("provided endpoint http://www.with space.url is invalid");
+    });
+    it("called with endpoint that is an invalid absolute url (invalid scheme) throws", () => {
+        expect(() => completeEndpoint("ht!ps://www.invalid-scheme.url")).toThrow("provided endpoint ht!ps://www.invalid-scheme.url is invalid");
+    });
+    it("called with endpoint that is an invalid absolute url (incomplete) throws", () => {
+        expect(() => completeEndpoint("https://")).toThrow("provided endpoint https:// is invalid");
+    });
+    it("called with endpoint that is a relative url without leading slash throws", () => {
+        expect(() => completeEndpoint("endpoint/without/leading/slash")).toThrow("provided endpoint endpoint/without/leading/slash is invalid");
     });
 });