+++ packages/server/src/transport/transport-request.spec.ts
@@ -6,7 +6,7 @@
  */
 
 import { serialize } from "@quatico/magellan-shared";
-import { addNamespace, addTransport, initProjectConfiguration } from "../configuration";
+import { initProjectConfiguration, setNamespace, setTransport } from "../configuration";
 import { initDependencyContext } from "../services";
 import { transportRequest } from "./transport-request";
 
@@ -14,8 +14,8 @@ const transportHandler = jest.fn();
 
 beforeAll(() => {
     initDependencyContext({ defaultTransportRequest: jest.fn(), defaultTransportHandler: transportHandler });
-    addTransport("test", transportHandler);
-    addNamespace("test", { endpoint: "/api", transport: "test" });
+    setTransport("test", transportHandler);
+    setNamespace("test", { endpoint: "/api", transport: "test" });
 });
 
 afterEach(() => {